refactor!: split trait impls for Expression into standalone files inside /traits
